Split up Configuration file into two: RedisConnectionPool & RedisConnection
